requirements.txt
Flask~=3.0.3
langchain-core~=0.2.0
langsmith>=0.1.17,<0.2.0
python-dotenv~=1.0.0
SQLAlchemy~=1.4.30

